+++ allauth/account/tests.py
@@ -536,6 +536,85 @@ class AccountTests(TestCase):
                 else
                 'The username and/or password you specified are not correct.')
 
+    @override_settings(
+        ACCOUNT_AUTHENTICATION_METHOD=app_settings.AuthenticationMethod.EMAIL,
+        ACCOUNT_LOGIN_ATTEMPTS_LIMIT=1)
+    def test_login_failed_attempts_exceeded_cleared_on_password_reset(self):
+        # Ensure that login attempts, once they hit the limit,
+        # can use the password reset mechanism to regain access.
+        user = get_user_model().objects.create(
+            username='john', email="john@example.org", is_active=True)
+        user.set_password('doe')
+        user.save()
+
+        EmailAddress.objects.create(
+            user=user,
+            email='john@example.com',
+            primary=True,
+            verified=True
+        )
+
+        resp = self.client.post(
+            reverse('account_login'),
+            {'login': user.email, 'password': 'bad'})
+        self.assertFormError(
+            resp,
+            'form',
+            None,
+            'The e-mail address and/or password you specified'
+            ' are not correct.')
+
+        resp = self.client.post(
+            reverse('account_login'),
+            {'login': user.email, 'password': 'bad'})
+        self.assertFormError(
+            resp,
+            'form',
+            None,
+            'Too many failed login attempts. Try again later.')
+
+        self.client.post(
+            reverse('account_reset_password'),
+            data={'email': user.email})
+
+        body = mail.outbox[0].body
+        self.assertGreater(body.find('https://'), 0)
+
+        # Extract URL for `password_reset_from_key` view and access it
+        url = body[body.find('/password/reset/'):].split()[0]
+        resp = self.client.get(url)
+        # Follow the redirect the actual password reset page with the key
+        # hidden.
+        url = resp.url
+        resp = self.client.get(url)
+        self.assertTemplateUsed(
+            resp,
+            'account/password_reset_from_key.%s' %
+            app_settings.TEMPLATE_EXTENSION)
+        self.assertFalse('token_fail' in resp.context_data)
+
+        new_password = 'newpass123'
+
+        # Reset the password
+        resp = self.client.post(url,
+                                {'password1': new_password,
+                                 'password2': new_password})
+        self.assertRedirects(resp,
+                             reverse('account_reset_password_from_key_done'))
+
+        # Check the new password is in effect
+        user = get_user_model().objects.get(pk=user.pk)
+        self.assertTrue(user.check_password(new_password))
+
+        resp = self.client.post(
+            reverse('account_login'),
+            {'login': user.email,
+             'password': new_password})
+
+        self.assertRedirects(resp,
+                             settings.LOGIN_REDIRECT_URL,
+                             fetch_redirect_response=False)
+
     def test_login_unverified_account_mandatory(self):
         """Tests login behavior when email verification is mandatory."""
         user = get_user_model().objects.create(username='john')
@@ -1272,3 +1351,39 @@ class TestCVE2019_19844(TestCase):
         data = {'email': 'mike@ıxample.org'}
         form = ResetPasswordForm(data)
         self.assertFalse(form.is_valid())
+
+
+class RequestAjaxTests(TestCase):
+
+    def _send_post_request(self, **kwargs):
+        return self.client.post(
+            reverse('account_signup'), {
+                'username': 'johndoe',
+                'email': 'john@example.org',
+                'email2': 'john@example.org',
+                'password1': 'johndoe',
+                'password2': 'johndoe'
+            },
+            **kwargs
+        )
+
+    def test_no_ajax_header(self):
+        resp = self._send_post_request()
+        self.assertEqual(302, resp.status_code)
+        self.assertRedirects(
+            resp, settings.LOGIN_REDIRECT_URL, fetch_redirect_response=False
+        )
+
+    def test_ajax_header_x_requested_with(self):
+        resp = self._send_post_request(
+            HTTP_X_REQUESTED_WITH='XMLHttpRequest'
+        )
+        self.assertEqual(200, resp.status_code)
+        self.assertEqual(settings.LOGIN_REDIRECT_URL, resp.json()['location'])
+
+    def test_ajax_header_http_accept(self):
+        resp = self._send_post_request(
+            HTTP_ACCEPT='application/json'
+        )
+        self.assertEqual(200, resp.status_code)
+        self.assertEqual(settings.LOGIN_REDIRECT_URL, resp.json()['location'])