+++ example/example/settings.py
@@ -174,15 +174,9 @@ AUTH_PASSWORD_VALIDATORS = [
     }
 ]
 
+ALLOWED_HOSTS = ['127.0.0.1', 'localhost']
 
 try:
-    from local_settings import *  # noqa
+    from .local_settings import *  # noqa
 except ImportError:
     pass
-ALLOWED_HOSTS = ['5d825be1.ngrok.io','127.0.0.1']
-
-SOCIALACCOUNT_PROVIDERS = {
-    'edx': {
-        'EDX_URL': "https://draft.navoica.pl",
-    }
-}