+++ ChangeLog.rst
@@ -1,9 +1,16 @@
-0.43.0 (2020-06-27)
+0.43.0
+******
 
 Note worthy changes
 -------------------
 
 - New translation: Slovenian.
+
+- If ``ACCOUNT_LOGIN_ATTEMPTS_LIMIT`` is set and the user successfully
+  resets their password, the timeout is cleared to allow immediate login.
+
+- New provider: Zoom, Stocktwits, Zoho
+
 
 0.42.0 (2020-05-24)
 *******************